Add presenter logic to refresh photo gallery data
The logic to refresh photo gallery data has been added to PhotoGalleryPresenter.
When user start typing in the search bar we want to refresh gallery data.
In order to achieve this behaviour the presenter needs to dispatch a background task that fetches new query data and cancel the previous one if still pending.
Presenter is also responsible to create the new viewModel to render whenever a new fetch request is fullfilled.
The PhotoGalleryViewModel has been created having in mind this purpose.
If there's an error during data fetch, this will be notified to PhotoGalleryViewController aswell.
The ModuleComposer view reference binding have fixed.