Add the photo fetch from url
The PhotoGalleryCollectionViewCell has been enhance with photo fetch.
The fetch flow starts when the Cell gets reused and appears on the screen.
The presenter will be called and a background thread will be dispatched. At this point the Interactor gets called in order to fulfil data request.
The fetched data will then be used by the Cell imageView.
If something goes wrong the returned data from presenter completion block will be nil. I decided to notify it to the user by showing the reddit post title and the img icon.
The photo collection imageUrl evaluation have been enhanced as follow:
- If the post is an "image" itself then get url directly
- Try to get the post thumbnail otherwise
- Try to get the header image otherwise