Add the photo details render logic
In order to the image details when a user select it from gallery i added an UIScrollView with pagination to PhotoDetailsViewController.
The scrollView content will be evaluated every time that this module will appear on screen.
Based on the number of photo shown in gallery at "selection time" the same number of PhotoDetailsView will be added to the scrollView.
Every PhotoDetailsView will have the screen size and its position will be evaluated based on photo index.
The PhotoDetailsView will be responsible to render the single photo details, for now i added the title label.
When the user select a photo from gallery UIScrollView content position changes based on the photo index.